template/FullStack/Vite(Frontend)+Flask(Backend)+Gemini(AI)/server/requirements.txt
Flask
flask-cors
pymongo
python-dotenv
gunicorn
requests
google-genai